+++ src/main/java/com/petassegang/addons/creative/ModCreativeTab.java
@@ -38,7 +38,7 @@ public final class ModCreativeTab {
         output.accept(ModItems.GANG_BADGE.get());
         output.accept(ModItems.CURSED_SNACK.get());
         output.accept(ModItems.LEVEL_ZERO_WALLPAPER.get());
-        output.accept(ModItems.LEVEL_ZERO_WALLPAPER_AGED.get());
+        output.accept(ModItems.LEVEL_ZERO_WALL_INSULATION.get());
         output.accept(ModItems.LEVEL_ZERO_DAMP_CARPET.get());
         output.accept(ModItems.LEVEL_ZERO_DAMP_CARPET_AGED.get());
         output.accept(ModItems.LEVEL_ZERO_CEILING_TILE.get());